feat(web): item-page-and-refactoring
