+++ web/src/components/HistoryDisplay/Header.tsx
