+++ web/src/components/InformationCard/Policies.tsx
